+++ payload-types.ts
@@ -11,6 +11,7 @@ export interface Config {
     users: User;
     pages: Page;
     media: Media;
+    features: Feature;
     'payload-preferences': PayloadPreference;
     'payload-migrations': PayloadMigration;
   };
@@ -26,6 +27,11 @@ export interface Config {
  */
 export interface User {
   id: string;
+  roles: ('super-admin' | 'admin' | 'editor' | 'viewer')[];
+  personalInformation?: {
+    firstName?: string | null;
+    lastName?: string | null;
+  };
   updatedAt: string;
   createdAt: string;
   email: string;
@@ -80,6 +86,17 @@ export interface Media {
   height?: number | null;
   focalX?: number | null;
   focalY?: number | null;
+}
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "features".
+ */
+export interface Feature {
+  id: string;
+  feature: string;
+  isEnabled: boolean;
+  updatedAt: string;
+  createdAt: string;
 }
 /**
  * This interface was referenced by `Config`'s JSON-Schema