+++ src/collections/Media/index.ts
@@ -1,0 +1,14 @@
+import { CollectionConfig } from 'payload/types'
+
+const Media: CollectionConfig = {
+  slug: 'media',
+  upload: true,
+  fields: [
+    {
+      name: 'text',
+      type: 'text',
+    },
+  ],
+}
+
+export default Media