+++ src/components/AfterNavLinks/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import AfterNavLinksClient from './index.client'
+
+const AfterNavLinks: React.FC = () => {
+  return <AfterNavLinksClient />
+}
+
+export default AfterNavLinks